Arranged into more OOP
